implemented fread_ra (untested)

+++ shared_libs/pred/predictor.hpp
@@ -6,7 +6,7 @@
 
 
 bool handle_open(int fd, const char *filename);
-int handle_read(int fd, off_t pos, size_t bytes);
+size_t handle_read(int fd, off_t pos, size_t bytes);
 int handle_write(int fd, off_t pos, size_t bytes);
 int handle_close(int fd);
 void print_readahead_time();
